Fix handling Instrumental parameter
I noticed that API calls to /suno/create with the optional--instrumental parameter were still sending the lyric prompt. According to the Suno API docs, for instrumental tracks, I should use gpt_description_prompt instead of prompt.

I made some changes so when the --instrumental argument is addedd, the prompt value is sent as gpt_description_prompt (a description of the instrumental music you want)
When --instrumental is NOT used, the prompt value is sent as prompt (the actual lyrics).